feat(converter): add container queries and shared unit data
