+++ __tests__/converter.test.tsx
@@ -1,4 +1,4 @@
-import { convertUnit } from '../components/apps/converter/UnitConverter';
+import { convertUnit } from '../components/apps/converter/unitData';
 import { render, fireEvent, screen, waitFor } from '@testing-library/react';
 import UnitConverter from '../components/apps/converter/UnitConverter';
